logic/requirements.txt
selenium
webdriver-manager
chromedriver-binary-auto
configparser